backend/requirements.txt
# Core web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.20
pydantic==2.12.0

# ML/NLP - Core
torch==2.4.0
transformers==4.44.2
openai-whisper==20250625

# ML/NLP - Supporting
numpy
tokenizers
huggingface-hub
safetensors
tqdm

# Training dependencies (only if needed)
datasets
scikit-learn

# Utils
python-dotenv
